requirements.txt
spotipy
beautifulsoup4
requests
python-dotenv